mvp_ai/requirements.txt
torch
torchvision
openai-clip
torchvision
transformers
requests
flask
open-clip-torch
flask 